+++ services/server/package.json
@@ -17,6 +17,7 @@
     "postgres-test:stop": "docker-compose -f test/docker-compose.yml rm -fsv",
     "postgres-test:migrate": "cd ../database && git submodule update --init && SERVICE_NAME=server DATABASE_URL=postgres://sourcify:sourcify@${SOURCIFY_POSTGRES_HOST:-sourcify-db}:${DOCKER_HOST_POSTGRES_TEST_PORT:-5432}/sourcify?sslmode=disable DBMATE_SCHEMA_FILE=./sourcify-database.sql npm run migrate:up-no-dump",
     "test:unit": "mocha  --exit --recursive test/unit/**",
+    "test:creation-tx-apis": "mocha --exit --recursive test/creation-tx-apis/**",
     "test": "npm run postgres-test:migrate & npx c8 --reporter=none mocha --exit --recursive test/unit/** test/integration/**",
     "test-local": "export DOCKER_HOST_POSTGRES_TEST_PORT=${DOCKER_HOST_POSTGRES_TEST_PORT:-5431} && npm run postgres-test:start && sleep 2 && SOURCIFY_POSTGRES_HOST=localhost npm run test; status=$?; npm run postgres-test:stop; exit $status",
     "generate-test-case": "ts-node test/integration/verification-cases/generate-test-case.ts",
